Доработка сервиса магазина(заказ, доставка, оплата).

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/interactionapi/warehouse/BookedProductsDto.java
@@ -3,13 +3,15 @@ package ru.yandex.practicum.commerce.interactionapi.warehouse;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class BookedProductsDto {
-    Double deliveryWeight;
-    Double deliveryVolume;
+    BigDecimal deliveryWeight;
+    BigDecimal deliveryVolume;
     Boolean fragile;
 }
